refactor: 메서드 명 변경
